Move tests/CHIPClusters.cpp to tests/CHIPClustersTests.cpp
This file has the same name as zap-generated/CHIPClusters.cpp, which
causes problems with some integrations and may cause confusion.
Rename it to CHIPClustersTest.cpp.
