+++ src/darwin/Framework/CHIP/templates/CHIPTestClustersObjc-src.zapt
@@ -7,7 +7,7 @@
 #import "CHIPDevice.h"
 #import "CHIPDevice_Internal.h"
 
-#import "zap-generated/tests/CHIPClusters.h"
+#import "zap-generated/tests/CHIPClustersTest.h"
 #import "zap-generated/CHIPTestClustersObjc.h"
 
 using chip::Callback::Cancelable;
