+++ zzz_generated/controller-clusters/zap-generated/tests/CHIPClustersTest.cpp
@@ -17,7 +17,7 @@
 
 // THIS FILE IS GENERATED BY ZAP
 
-#include "CHIPClusters.h"
+#include <zap-generated/tests/CHIPClustersTest.h>
 
 #include <app/InteractionModelEngine.h>
 #include <app/util/basic-types.h>
